backend/requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Audio processing
PyDejavu==1.0.0
mysql-connector-python==8.2.0

# HTTP requests (for GurbaniNow API proxy)
requests==2.31.0

# Utilities
Werkzeug==3.0.1

